added all on table CandidateEvent

+++ sandbox-web/src/main/java/com/exadel/team2/sandbox/controller/CandidateEventController.java
@@ -1,11 +1,14 @@
 package com.exadel.team2.sandbox.controller;
 
-import com.exadel.team2.sandbox.entity.CandidateEventEntity;
 import com.exadel.team2.sandbox.service.CandidateEventService;
+import com.exadel.team2.sandbox.web.candidate_event.CreateCandidateEventDto;
+import com.exadel.team2.sandbox.web.candidate_event.ResponseCandidateEventDto;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/candidate-event")
@@ -14,28 +17,28 @@ public class CandidateEventController {
     private final CandidateEventService candidateEventService;
 
     @GetMapping("/{id}")
-    public CandidateEventEntity getCandidateEvent(@PathVariable Long id) {
+    public ResponseCandidateEventDto getCandidateEvent(@PathVariable Long id) {
         return candidateEventService.getById(id);
     }
 
     @GetMapping("/all")
-    public List<CandidateEventEntity> getAllCandidateEvent() {
-        return candidateEventService.getAll();
+    public Page<ResponseCandidateEventDto> getAllCandidateEvent(
+            @RequestParam(defaultValue = "0", name = "page") Integer page,
+            @RequestParam(defaultValue = "15", name = "itemsPerPage") Integer itemsPerPage) {
+        return candidateEventService.getAllPageable(PageRequest.of(page, itemsPerPage));
     }
 
     @PostMapping
-    public CandidateEventEntity saveCandidateEvent(@RequestBody CandidateEventEntity entity) {
-        return candidateEventService.save(entity);
-    }
-
-    @PutMapping("/{id}")
-    public CandidateEventEntity updateCandidateEvent(@PathVariable Long id, @RequestBody CandidateEventEntity entity) {
-        entity.setId(id);
-        return candidateEventService.update(entity);
+    public ResponseCandidateEventDto createCandidateEvent(@Validated @RequestBody CreateCandidateEventDto createCandidateEventDto) {
+        return candidateEventService.save(createCandidateEventDto);
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id) {
+    public ResponseEntity<?> deleteCandidateEvent(@PathVariable Long id) {
+        if (candidateEventService.getById(id) == null) {
+            return ResponseEntity.notFound().build();
+        }
         candidateEventService.delete(id);
+        return ResponseEntity.ok().build();
     }
 }
